move PiecePositionPointsBuilder to builders
