+++ include/builders.hpp
@@ -2,7 +2,7 @@
 
 #include <game_board_for_concepts.hpp>
 #include <memory>
-
+#include <piece_position_points_for_concepts.hpp>
 namespace gameboard {
 
 class GameBoardBuilder {
@@ -12,5 +12,27 @@ public:
   );
 };
 
+} // namespace gameboard
+
+
+namespace piecepoints {
+    class PiecePositionPointsBuilder {
+public:
+  std::shared_ptr<PiecePositionPointsForConcepts> build(
+      std::string json_file = piecepoints::kICGABPOPath
+  );
+
+  // Don't need any of these .build() overloads for now, but declare them so we remember
+  // that we might want to implement them at some point.
+  std::shared_ptr<PiecePositionPointsForConcepts> build(
+      GamePointsArray_t game_points_array
+  );
+  std::shared_ptr<PiecePositionPointsForConcepts> build(BPOPointsEKeys &bpo_points_ekeys
+  );
+  std::shared_ptr<PiecePositionPointsForConcepts> build(BPOPointsSKeys &bpo_points_skeys
+  );
+};
 }
+
+
